% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assert_scalar_numeric}
\alias{.assert_scalar_numeric}
\title{Assert that a scalar is numeric}
\usage{
.assert_scalar_numeric(x, nm = deparse(substitute(x)))
}
\arguments{
\item{x}{Object to test.}

\item{nm}{Character name used in the error message.}
}
\description{
Thin wrapper around \code{stopifnot()} providing a clearer error message when an
argument is expected to be a single numeric value.
}
\keyword{internal}
