% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sanitize_weights}
\alias{.sanitize_weights}
\title{Sanitise a weight vector (internal)}
\usage{
.sanitize_weights(w)
}
\description{
Replaces NA with 0 and clips machine-precision negatives to 0.
}
\keyword{internal}
